backend/algorand/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
openai
python-dotenv
google-genai
ollama
boto3
botocore
requests
paramiko
cryptography
langgraph
langchain-core
typing-extensions
pydantic
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
httpx
bcrypt
py-algorand-sdk
fastapi
uvicorn
python-dotenv
cryptography
py-algorand-sdk
algosdk
pydantic
algorand-sdk
puyapy
